Moved the logic for product styles higher in the component stream.
