+++ client/src/components/ProductDetailPage.jsx
@@ -6,15 +6,12 @@ import MainImageGallery from './MainImageGallery.jsx';
 import ProductDescription from './ProductDescription.jsx';
 
 const ProductDetailPageComponent = (props) => {
-  const [productDetails, setProductDetails] = useState({});
   const { productId } = useParams();
+  const [productDetails, setProductDetails] = useState({});
+  const [styles, setStyles] = useState([]);
+  const [selectedStyle, setSelectedStyle] = useState({})
 
-
-  useEffect(() => {
-    getProductDetails();
-  },[])
-
-  let getProductDetails = () => {
+  const getProductDetails = () => {
     axios.get(`/api/products/${productId}`)
     .then((data) => {
       setProductDetails(data.data)
@@ -24,16 +21,34 @@ const ProductDetailPageComponent = (props) => {
     })
   }
 
-  let styles = {
-    "display": "flex",
+  const getStyles = () => {
+    axios.get(`/api/products/${productId}/styles`)
+      .then(res => {
+        setStyles(res.data.results);
+        setSelectedStyle(res.data.results[0]);
+      })
+      .catch(err => {
+        console.log(err);
+      });
+  }
 
+  useEffect(()=> {
+    getStyles();
+  }, [productDetails])
+
+  useEffect(() => {
+    getProductDetails();
+  },[])
+
+  let productDetailStyles = {
+    "display": "flex",
   }
 
   return (
     <div>
-    <div style={styles}>
+    <div style={productDetailStyles}>
       <MainImageGallery />
-      <ProductDescription productDetails={productDetails} productId={productId}/>
+      <ProductDescription productDetails={productDetails} productId={productId} styles={styles} selectedStyle={selectedStyle}/>
     </div>
     {productDetails.description}
     </div>
